+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+- Add support for custom headers in the pytest web server code, by setting
+  the `extra_headers` parameter in the `spawn_web_server` function.
+  [#39](https://github.com/pyodide/pytest-pyodide/pull/39)
+
 - Breaking: removed STANDALONE_REFRESH env variable which was used to
   override `selenium_standalone` fixture with `selenium_standalone_refresh`.
   [#65](https://github.com/pyodide/pytest-pyodide/pull/65)